requirements.txt
# Data Analysis
pandas>=1.5.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
textstat>=0.7.3
wordcloud>=1.9.2

# Existing dependencies
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
google-generativeai==0.3.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pip==22.3.1
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.5.2
pydantic-core==2.14.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
numpy<2
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.1
torch
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
